+++ lib/doggo/components/toolbar.ex
@@ -65,7 +65,7 @@ defmodule Doggo.Components.Toolbar do
   @impl true
   def config do
     [
-      type: :miscellaneous,
+      type: :buttons,
       since: "0.6.0",
       maturity: :developing,
       modifiers: []